@@ -37,40 +37,44 @@
     "onCommand:PowerShell.ShowSessionConsole",
     "onCommand:PowerShell.ShowSessionMenu",
     "onCommand:PowerShell.RestartSession",
+    "onCommand:PowerShell.ShowLogs",
+    "onCommand:PowerShell.OpenLogFolder",
+    "onCommand:PowerShell.GenerateBugReport",
+    "onCommand:PowerShell.OpenExamplesFolder",
     "onCommand:PowerShell.EnableISEMode",
     "onCommand:PowerShell.DisableISEMode",
     "onView:PowerShellCommands"
   ],
   "dependencies": {
+    "@vscode/extension-telemetry": "~0.6.2",
     "node-fetch": "~2.6.7",
-    "semver": "~7.3.5",
+    "semver": "~7.3.7",
     "uuid": "~8.3.2",
-    "vscode-extension-telemetry": "~0.4.5",
     "vscode-languageclient": "~7.0.0",
-    "vscode-languageserver-protocol": "~3.16.0"
+    "vscode-languageserver-protocol": "~3.17.2"
   },
   "devDependencies": {
     "@types/glob": "~7.2.0",
-    "@types/mocha": "~9.1.0",
+    "@types/mocha": "~9.1.1",
     "@types/mock-fs": "~4.13.1",
     "@types/node": "~14.17.4",
     "@types/node-fetch": "~2.6.1",
     "@types/rewire": "~2.5.28",
-    "@types/semver": "~7.3.9",
-    "@types/sinon": "~10.0.11",
+    "@types/semver": "~7.3.10",
+    "@types/sinon": "~10.0.13",
     "@types/uuid": "~8.3.4",
-    "@types/vscode": "~1.56.0",
-    "@vscode/test-electron": "~2.1.2",
-    "esbuild": "^0.14.23",
-    "glob": "^7.2.0",
-    "mocha": "~9.2.1",
+    "@types/vscode": "~1.59.0",
+    "@vscode/test-electron": "~2.1.5",
+    "esbuild": "^0.14.51",
+    "glob": "^8.0.3",
+    "mocha": "~10.0.0",
     "mocha-multi-reporters": "~1.5.1",
     "mock-fs": "~5.1.2",
     "rewire": "~6.0.0",
-    "sinon": "~13.0.1",
+    "sinon": "~14.0.0",
     "tslint": "~6.1.3",
-    "typescript": "~4.5.5",
-    "vsce": "~2.6.7"
+    "typescript": "~4.7.4",
+    "vsce": "~2.10.0"
   },
   "extensionDependencies": [
     "vscode.powershell"
