@@ -8,10 +8,8 @@ import { NotificationType, RequestType } from "vscode-languageclient";
 import { LanguageClient } from "vscode-languageclient/node";
 import { getPlatformDetails, OperatingSystem } from "../platform";
 import { PowerShellProcess} from "../process";
-import { SessionManager, SessionStatus } from "../session";
+import { IEditorServicesSessionDetails, SessionManager, SessionStatus } from "../session";
 import Settings = require("../settings");
-import utils = require("../utils");
-import { NamedPipeDebugAdapter } from "../debugAdapter";
 import { Logger } from "../logging";
 import { LanguageClientConsumer } from "../languageClientConsumer";
 
@@ -26,7 +24,7 @@ export class DebugSessionFeature extends LanguageClientConsumer
 
     private sessionCount: number = 1;
     private tempDebugProcess: PowerShellProcess;
-    private tempSessionDetails: utils.IEditorServicesSessionDetails;
+    private tempSessionDetails: IEditorServicesSessionDetails;
 
     constructor(context: ExtensionContext, private sessionManager: SessionManager, private logger: Logger) {
         super();
@@ -37,20 +35,16 @@ export class DebugSessionFeature extends LanguageClientConsumer
 
     createDebugAdapterDescriptor(
         session: vscode.DebugSession,
-        _executable: vscode.DebugAdapterExecutable): vscode.ProviderResult<vscode.DebugAdapterDescriptor> {
+        _executable: vscode.DebugAdapterExecutable | undefined): vscode.ProviderResult<vscode.DebugAdapterDescriptor> {
 
         const sessionDetails = session.configuration.createTemporaryIntegratedConsole
             ? this.tempSessionDetails
             : this.sessionManager.getSessionDetails();
 
-        // Establish connection before setting up the session
         this.logger.writeVerbose(`Connecting to pipe: ${sessionDetails.debugServicePipeName}`);
         this.logger.writeVerbose(`Debug configuration: ${JSON.stringify(session.configuration)}`);
 
-        const debugAdapter = new NamedPipeDebugAdapter(sessionDetails.debugServicePipeName, this.logger);
-        debugAdapter.start();
-
-        return new vscode.DebugAdapterInlineImplementation(debugAdapter);
+        return new vscode.DebugAdapterNamedPipeServer(sessionDetails.debugServicePipeName);
     }
 
     // tslint:disable-next-line:no-empty
@@ -60,19 +54,16 @@ export class DebugSessionFeature extends LanguageClientConsumer
     public setLanguageClient(languageClient: LanguageClient) {
         languageClient.onNotification(
             StartDebuggerNotificationType,
-            () =>
-                // TODO: Use a named debug configuration.
-                vscode.debug.startDebugging(undefined, {
-                    request: "launch",
-                    type: "PowerShell",
-                    name: "PowerShell: Interactive Session",
-        }));
+            // TODO: Use a named debug configuration.
+            () => vscode.debug.startDebugging(undefined, {
+                request: "launch",
+                type: "PowerShell",
+                name: "PowerShell: Interactive Session"
+            }));
 
         languageClient.onNotification(
             StopDebuggerNotificationType,
-            () =>
-                vscode.debug.stopDebugging(undefined)
-        );
+            () => vscode.debug.stopDebugging(undefined));
     }
 
     public async provideDebugConfigurations(
@@ -100,7 +91,7 @@ export class DebugSessionFeature extends LanguageClientConsumer
             {
                 id: DebugConfig.InteractiveSession,
                 label: "Interactive Session",
-                description: "Debug commands executed from the Integrated Console",
+                description: "Debug commands executed from the PowerShell Extension Terminal",
             },
             {
                 id: DebugConfig.AttachHostProcess,
@@ -239,7 +230,7 @@ export class DebugSessionFeature extends LanguageClientConsumer
                         : currentDocument.fileName;
 
             } else {
-                // If the non-temp integrated console is being used, default to the current working dir.
+                // If the non-temp Extension Terminal is being used, default to the current working dir.
                 config.cwd = "";
             }
         }
@@ -318,23 +309,9 @@ export class DebugSessionFeature extends LanguageClientConsumer
         // Create or show the interactive console
         vscode.commands.executeCommand("PowerShell.ShowSessionConsole", true);
 
-        const sessionFilePath = utils.getDebugSessionFilePath();
-
         if (config.createTemporaryIntegratedConsole) {
-            if (this.tempDebugProcess) {
-                this.tempDebugProcess.dispose();
-            }
-
-            this.tempDebugProcess =
-                this.sessionManager.createDebugSessionProcess(
-                    sessionFilePath,
-                    settings);
-
+            this.tempDebugProcess = this.sessionManager.createDebugSessionProcess(settings);
             this.tempSessionDetails = await this.tempDebugProcess.start(`DebugSession-${this.sessionCount++}`);
-            utils.writeSessionFile(sessionFilePath, this.tempSessionDetails);
-
-        } else {
-            utils.writeSessionFile(sessionFilePath, this.sessionManager.getSessionDetails());
         }
 
         return config;
@@ -374,7 +351,7 @@ export class SpecifyScriptArgsFeature implements vscode.Disposable {
 
         const text = await vscode.window.showInputBox(options);
         // When user cancel's the input box (by pressing Esc), the text value is undefined.
-        // Let's not blow away the previous settting.
+        // Let's not blow away the previous setting.
         if (text !== undefined) {
             this.context.workspaceState.update(powerShellDbgScriptArgsKey, text);
         }
@@ -473,7 +450,7 @@ export class PickPSHostProcessFeature extends LanguageClientConsumer {
         // Start with the current PowerShell process in the list.
         const items: IProcessItem[] = [{
             label: "Current",
-            description: "The current PowerShell Integrated Console process.",
+            description: "The current PowerShell Extension process.",
             pid: "current",
         }];
         for (const p in hostProcesses) {
