@@ -10,19 +10,6 @@ import vscode = require("vscode");
 
 export const PowerShellLanguageId = "powershell";
 
-export function ensurePathExists(targetPath: string): void {
-    // Ensure that the path exists
-    try {
-        // TODO: Use vscode.workspace.fs
-        fs.mkdirSync(targetPath);
-    } catch (e) {
-        // If the exception isn't to indicate that the folder exists already, rethrow it.
-        if (e.code !== "EEXIST") {
-            throw e;
-        }
-    }
-}
-
 // Check that the file exists in an asynchronous manner that relies solely on the VS Code API, not Node's fs library.
 export async function fileExists(targetPath: string | vscode.Uri): Promise<boolean> {
     try {
@@ -50,59 +37,19 @@ export function getPipePath(pipeName: string) {
     }
 }
 
-export interface IEditorServicesSessionDetails {
-    status: string;
-    reason: string;
-    detail: string;
-    powerShellVersion: string;
-    channel: string;
-    languageServicePort: number;
-    debugServicePort: number;
-    languageServicePipeName: string;
-    debugServicePipeName: string;
-}
-
-export type IReadSessionFileCallback = (details: IEditorServicesSessionDetails) => void;
-
-const sessionsFolder = path.resolve(__dirname, "../sessions");
-const sessionFilePathPrefix = path.resolve(sessionsFolder, "PSES-VSCode-" + process.env.VSCODE_PID);
-
-// Create the sessions path if it doesn't exist already
-ensurePathExists(sessionsFolder);
-
-export function getSessionFilePath(uniqueId: number) {
-    return `${sessionFilePathPrefix}-${uniqueId}`;
-}
-
-export function getDebugSessionFilePath() {
-    return `${sessionFilePathPrefix}-Debug`;
-}
-
-export function writeSessionFile(sessionFilePath: string, sessionDetails: IEditorServicesSessionDetails) {
-    ensurePathExists(sessionsFolder);
-
-    const writeStream = fs.createWriteStream(sessionFilePath);
-    writeStream.write(JSON.stringify(sessionDetails));
-    writeStream.close();
-}
-
-
-export function readSessionFile(sessionFilePath: string): IEditorServicesSessionDetails {
-    const fileContents = fs.readFileSync(sessionFilePath, "utf-8");
-    return JSON.parse(fileContents);
-}
-
-export function deleteSessionFile(sessionFilePath: string) {
-    try {
-        fs.unlinkSync(sessionFilePath);
-    } catch (e) {
-        // TODO: Be more specific about what we're catching
-    }
-}
-
 export function checkIfFileExists(filePath: string): boolean {
     try {
         fs.accessSync(filePath, fs.constants.R_OK);
+        return true;
+    } catch (e) {
+        return false;
+    }
+}
+
+export function checkIfDirectoryExists(directoryPath: string): boolean {
+    try {
+        // tslint:disable-next-line:no-bitwise
+        fs.accessSync(directoryPath, fs.constants.R_OK | fs.constants.O_DIRECTORY);
         return true;
     } catch (e) {
         return false;
